se corrigen rutas

+++ app/Http/routes.php
@@ -83,28 +83,28 @@ Route::get('callback', 'Admin\FacebookController@callback');
 
 
 
-Route::get('admin/tokens', ['as'=> 'admin.tokens.index', 'uses' => 'Admin\TokenController@index']);
-Route::post('admin/tokens', ['as'=> 'admin.tokens.store', 'uses' => 'Admin\TokenController@store']);
-Route::get('admin/tokens/create', ['as'=> 'admin.tokens.create', 'uses' => 'Admin\TokenController@create']);
-Route::put('admin/tokens/{tokens}', ['as'=> 'admin.tokens.update', 'uses' => 'Admin\TokenController@update']);
-Route::patch('admin/tokens/{tokens}', ['as'=> 'admin.tokens.update', 'uses' => 'Admin\TokenController@update']);
-Route::delete('admin/tokens/{tokens}', ['as'=> 'admin.tokens.destroy', 'uses' => 'Admin\TokenController@destroy']);
-Route::get('admin/tokens/{tokens}', ['as'=> 'admin.tokens.show', 'uses' => 'Admin\TokenController@show']);
-Route::get('admin/tokens/{tokens}/edit', ['as'=> 'admin.tokens.edit', 'uses' => 'Admin\TokenController@edit']);
+Route::get('admin/tokens', ['as'=> 'admin.tokens.index', 'uses' => 'Admin\TokenController@index', 'middleware' => ['auth']]);
+Route::post('admin/tokens', ['as'=> 'admin.tokens.store', 'uses' => 'Admin\TokenController@store', 'middleware' => ['auth']]);
+Route::get('admin/tokens/create', ['as'=> 'admin.tokens.create', 'uses' => 'Admin\TokenController@create', 'middleware' => ['auth']]);
+Route::put('admin/tokens/{tokens}', ['as'=> 'admin.tokens.update', 'uses' => 'Admin\TokenController@update', 'middleware' => ['auth']]);
+Route::patch('admin/tokens/{tokens}', ['as'=> 'admin.tokens.update', 'uses' => 'Admin\TokenController@update', 'middleware' => ['auth']]);
+Route::delete('admin/tokens/{tokens}', ['as'=> 'admin.tokens.destroy', 'uses' => 'Admin\TokenController@destroy', 'middleware' => ['auth']]);
+Route::get('admin/tokens/{tokens}', ['as'=> 'admin.tokens.show', 'uses' => 'Admin\TokenController@show', 'middleware' => ['auth']]);
+Route::get('admin/tokens/{tokens}/edit', ['as'=> 'admin.tokens.edit', 'uses' => 'Admin\TokenController@edit', 'middleware' => ['auth']]);
 
 
-Route::get('admin/formularios', ['as'=> 'admin.formularios.index', 'uses' => 'Admin\FormularioController@index']);
-Route::post('admin/formularios', ['as'=> 'admin.formularios.store', 'uses' => 'Admin\FormularioController@store']);
-Route::get('admin/formularios/create', ['as'=> 'admin.formularios.create', 'uses' => 'Admin\FormularioController@create']);
-Route::put('admin/formularios/{formularios}', ['as'=> 'admin.formularios.update', 'uses' => 'Admin\FormularioController@update']);
-Route::patch('admin/formularios/{formularios}', ['as'=> 'admin.formularios.update', 'uses' => 'Admin\FormularioController@update']);
-Route::delete('admin/formularios/{formularios}', ['as'=> 'admin.formularios.destroy', 'uses' => 'Admin\FormularioController@destroy']);
-Route::get('admin/formularios/{formularios}', ['as'=> 'admin.formularios.show', 'uses' => 'Admin\FormularioController@show']);
-Route::get('admin/formularios/{formularios}/edit', ['as'=> 'admin.formularios.edit', 'uses' => 'Admin\FormularioController@edit']);
-Route::get('admin/formularios/{formularios}/crear_estructura', ['as'=> 'admin.formularios.crear_estructura', 'uses' => 'Admin\FormularioController@crearEstructura']);
+Route::get('admin/formularios', ['as'=> 'admin.formularios.index', 'uses' => 'Admin\FormularioController@index', 'middleware' => ['auth']]);
+Route::post('admin/formularios', ['as'=> 'admin.formularios.store', 'uses' => 'Admin\FormularioController@store', 'middleware' => ['auth']]);
+Route::get('admin/formularios/create', ['as'=> 'admin.formularios.create', 'uses' => 'Admin\FormularioController@create', 'middleware' => ['auth']]);
+Route::put('admin/formularios/{formularios}', ['as'=> 'admin.formularios.update', 'uses' => 'Admin\FormularioController@update', 'middleware' => ['auth']]);
+Route::patch('admin/formularios/{formularios}', ['as'=> 'admin.formularios.update', 'uses' => 'Admin\FormularioController@update', 'middleware' => ['auth']]);
+Route::delete('admin/formularios/{formularios}', ['as'=> 'admin.formularios.destroy', 'uses' => 'Admin\FormularioController@destroy', 'middleware' => ['auth']]);
+Route::get('admin/formularios/{formularios}', ['as'=> 'admin.formularios.show', 'uses' => 'Admin\FormularioController@show', 'middleware' => ['auth']]);
+Route::get('admin/formularios/{formularios}/edit', ['as'=> 'admin.formularios.edit', 'uses' => 'Admin\FormularioController@edit', 'middleware' => ['auth']]);
+Route::get('admin/formularios/{formularios}/crear_estructura', ['as'=> 'admin.formularios.crear_estructura', 'uses' => 'Admin\FormularioController@crearEstructura', 'middleware' => ['auth']]);
 
 
-Route::get('admin/formularios/{formularios}/listar_datos', ['as'=> 'admin.formularios.listar_datos', 'uses' => 'Admin\FormularioController@listarDatos']);
+Route::get('admin/formularios/{formularios}/listar_datos', ['as'=> 'admin.formularios.listar_datos', 'uses' => 'Admin\FormularioController@listarDatos', 'middleware' => ['auth']]);
